requirements.txt
python-telegram-bot[job-queue]==21.6
psycopg==3.2.1
python-dotenv